+++ src/tpysui/widgets/editable_table.py
@@ -134,7 +134,7 @@ class CellConfig:
 class EditableDataTable(DataTable):
     """A datatable where you can edit cells."""
 
-    BINDINGS = [("e", "edit", "Edit Cell")]
+    BINDINGS = [("ctrl+e", "edit", "Edit Cell")]
 
     class CellValueChange(Message):
 
@@ -186,21 +186,22 @@ class EditableDataTable(DataTable):
             self.post_message(self.RowDelete(event.data_table, event.cell_key.row_key))
 
     async def action_edit(self) -> None:
-        coords = self.cursor_coordinate
-        # Avoid the 'delete' button
-        if coords.column != self.delete_column:
-            edit_cfg = self.edit_config[coords.column]
-            if edit_cfg.editable == False:
-                return
-            if edit_cfg.field_name == "Active" and self.row_count == 1:
-                self.app.push_screen(
-                    OkPopup("[red]Can not change Active state for only row")
-                )
-            else:
-                if edit_cfg.inline:
-                    self.edit_cell(coordinate=coords, cfg=edit_cfg)
-                elif edit_cfg.dialog:
-                    self.edit_dialog(coordinate=coords, cfg=edit_cfg)
+        if self.ordered_rows:
+            coords = self.cursor_coordinate
+            # Avoid the 'delete' button
+            if coords.column != self.delete_column:
+                edit_cfg = self.edit_config[coords.column]
+                if edit_cfg.editable == False:
+                    return
+                if edit_cfg.field_name == "Active" and self.row_count == 1:
+                    self.app.push_screen(
+                        OkPopup("[red]Can not change Active state for only row")
+                    )
+                else:
+                    if edit_cfg.inline:
+                        self.edit_cell(coordinate=coords, cfg=edit_cfg)
+                    elif edit_cfg.dialog:
+                        self.edit_dialog(coordinate=coords, cfg=edit_cfg)
 
     @work()
     async def edit_dialog(self, coordinate: Coordinate, cfg: CellConfig) -> None: